requirements.txt
django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-constance==4.3.2
PyJWT==2.10.1
psycopg2==2.9.10
redis==5.2.1
